+++ Modules/Changelog.lua
@@ -4,6 +4,14 @@ local changelogFrame
 
 local CHANGELOG_ENTRIES = {
 	{
+		version = "2.2.0-beta.1",
+		lines = {
+			"CHANGELOG_220_1",
+			"CHANGELOG_220_2",
+			"CHANGELOG_220_3",
+		},
+	},
+	{
 		version = "2.1.1",
 		lines = {
 			"CHANGELOG_211_1",